backend/requirements.txt
google-adk>=1.0.0,<2.0.0
google-genai>=1.0.0,<2.0.0
google-cloud-logging>=3.10.0,<4.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
firebase-admin>=6.0.0,<7.0.0
tenacity>=8.2.0,<10.0.0
